comments, cleanup

+++ N/src/n/graphics/sprite_atlas_bank.h
@@ -5,12 +5,12 @@ namespace n
 {
 	class sprite_atlas;
 
-	class sprite_atlas_bank final : public mage::serializable_bank<sprite_atlas, uint8_t, c::sprite_atlas_bank_size>
+	class sprite_atlas_bank final : public mage::serializable_bank<sprite_atlas, uint16_t, c::sprite_atlas_bank_size>
 	{
 	public:
 		sprite_atlas_bank() {}
 		sprite_atlas_bank(mage::input_file& in) :
-			mage::serializable_bank<sprite_atlas, uint8_t, c::sprite_atlas_bank_size>(in)
+			mage::serializable_bank<sprite_atlas, uint16_t, c::sprite_atlas_bank_size>(in)
 		{}
 		N_DCM(sprite_atlas_bank);
 	};
